requirements.txt
opencv-python==4.8.0.76
Pillow==10.0.0
pymongo==4.5.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
requests==2.31.0
